refactor(websites): extraer helpers de DesignPanel y editor page
- Crear design-panel-helpers.tsx con constantes (FONT_PAIRINGS,
  STYLE_OPTIONS, PRESET_COLORS, etc), color utilities (hexToHsl,
  hslToHex, generateHarmonies), y sub-componentes (StyleMockup,
  Section, ColorPicker, FontDropdown)
- Crear editor/_helpers.ts con tipos (SectionContent, ChatMsg,
  ThemeData, ActiveTab) y constantes (DEFAULT_THEME, TOP_BAR_HEIGHT)
- Reducir DesignPanel.tsx de 970 a 556 lineas (-414)
- Reducir editor/page.tsx de 1693 a 1643 lineas (-50)
- Eliminar 3 warnings de unused imports en DesignPanel

Refs: #244
